requirements.txt
# Core Framework
Flask==2.3.2
Waitress==2.1.2

# Database
pyodbc==4.0.39

# Authentication
ldap3==2.9.1
python-ldap==3.4.3

# Environment
python-dotenv==1.0.0

# Internationalization (NEW)
Flask-Babel==4.0.0
Babel==2.13.1

# Security
Werkzeug==2.3.6
MarkupSafe==2.1.3
Jinja2==3.1.2
cryptography==41.0.4

# Excel Export (NEW)
openpyxl==3.1.2